% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solar-time.r
\name{format.solar_time}
\alias{format.solar_time}
\title{Encode in a Common Format}
\usage{
\method{format}{solar_time}(x, ..., sep = ":")
}
\arguments{
\item{x}{an R object}

\item{...}{ignored}

\item{sep}{character used as separator}
}
\value{
A character string with the time formatted as "HH:MM:SS", where ":"
is the argument passed to \code{sep}.
}
\description{
Format a \code{solar_time} object for pretty printing
}
\seealso{
Other astronomy related functions: 
\code{\link{day_night}()},
\code{\link{sun_angles}()}
}
\concept{astronomy related functions}
