% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time-of-day.r
\name{format.tod_time}
\alias{format.tod_time}
\title{Encode in a Common Format}
\usage{
\method{format}{tod_time}(x, ..., sep = ":")
}
\arguments{
\item{x}{an R object}

\item{...}{ignored}

\item{sep}{character used as separator}
}
\value{
A character string with the time formatted as "HH:MM:SS", where ":"
is the argument passed to \code{sep}.
}
\description{
Format a \code{tod_time} object into a character string for pretty printing.
}
\seealso{
Other Time of day functions: 
\code{\link{as_tod}()},
\code{\link{print.tod_time}()}
}
\concept{Time of day functions}
