% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/photobiologySunCalc.r
\docType{package}
\name{photobiologySunCalc-package}
\alias{photobiologySunCalc}
\alias{photobiologySunCalc-package}
\title{photobiologySunCalc: Sun and Atmosphere Calculations}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Compute the position of the sun, day and night length, local solar time using Meeus' very accurate formulae. Estimate air mass (AM) from solar elevation, reference evapotranspiration, and interconvert air water content expressed as different physical quantities.
}
\details{
Please see the vignette \emph{0: The R for
  Photobiology Suite} for a description of the suite.
}
\examples{
# daylength
sunrise_time(lubridate::today(tzone = "EET"), tz = "EET",
             geocode = data.frame(lat = 60, lon = 25),
             unit.out = "hour")
day_length(lubridate::today(tzone = "EET"), tz = "EET",
           geocode = data.frame(lat = 60, lon = 25),
           unit.out = "hour")
sun_angles(lubridate::now(tzone = "EET"), tz = "EET",
           geocode = data.frame(lat = 60, lon = 25))

water_vp_sat(23) # 23 C -> vapour pressure in Pa

}
\references{
Aphalo, Pedro J. (2015) The r4photobiology suite. \emph{UV4Plants Bulletin}, 2015:1,
21-29. \doi{10.19232/uv4pb.2015.1.14}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://docs.r4photobiology.info/xx/}
  \item \url{https://github.com/aphalo/xx}
  \item Report bugs at \url{https://github.com/aphalo/xx/issues}
}

}
\author{
\strong{Maintainer}: Pedro J. Aphalo \email{pedro.aphalo@helsinki.fi} (\href{https://orcid.org/0000-0003-3385-972X}{ORCID})

}
